Refactor : remove LadderGenerator
